add exceptions

+++ opinions_app/api_views.py
@@ -5,13 +5,16 @@ from flask import jsonify, request
 from random import randrange
 
 from . import app, db
+from .error_handlers import InvalidAPIUsage
 from .models import Opinion
 from .views import random_opinion
 
 
 @app.route('/api/opinions/<int:id>/', methods=['GET'])
 def get_opinion(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     return jsonify({'opinion': opinion.to_dict()}), 200
 
 
@@ -21,8 +24,10 @@ def update_opinion(id):
     if (
         'text' in data and
         Opinion.query.by_filter(text=data['text']).first() is not None):
-        return jsonify({'error': 'Такое мнение уже есть в базе данных'}), 400
-    opinion = Opinion.query.get_or_404(id)
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных')
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     # Если метод get_or_404 не найдёт указанный ключ,
     # то он выбросит исключение 404
     opinion.title = data.get('title', opinion.title)
@@ -37,7 +42,9 @@ def update_opinion(id):
 
 @app.route('/api/opinions/<int:id>/', methods=['DELETE'])
 def delete_opinin(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     db.session.delete(opinion)
     db.session.commit()
     return '', 204
@@ -53,9 +60,9 @@ def get_opinions():
 def add_opinion():
     data = request.get_json()
     if 'title' not in data or 'text' not in data:
-        return jsonify({'error': 'В запросе отсутствуют обязательные поля'}), 400
+        raise InvalidAPIUsage('В запросе отсутствуют обязательные поля')
     if Opinion.query.filter_by(text=data['text']).first() is not None:
-        return jsonify({"error": 'Такое мнение уже есть в базе данных'}), 400
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных')
     opinion = Opinion()
     opinion.from_dict(data)
     db.session.add(opinion)
@@ -65,4 +72,6 @@ def add_opinion():
 @app.route('/api/get-random-opinion/', methods=['GET'])
 def get_random_opinion():
     opinion = random_opinion()
-    return jsonify({'opinion': opinion.to_dict()}), 200
+    if opinion is not None:
+        return jsonify({'opinion': opinion.to_dict()}), 200
+    raise InvalidAPIUsage('В базе данных нет мнений', 404)
